+++ src/Plugins/PlaneMaterial/PlaneMaterial.hpp
@@ -7,27 +7,27 @@ namespace Plugin
     class PlaneMaterial : public Raytracer::AMaterial
     {
     public:
-        PlaneMaterial(const Component::Color &color1, const Component::Color &color2, float squareSize, float diffuseFactor, float reflectivity);
+        PlaneMaterial(const Component::Color &color1, const Component::Color &color2, double squareSize, double diffuseFactor, double reflectivity);
 
         ~PlaneMaterial() override = default;
 
         [[nodiscard]] Component::Color getDiffuse() const override;
 
-        [[nodiscard]] float getSpecular() const override;
+        [[nodiscard]] double getSpecular() const override;
 
-        [[nodiscard]] float getShininess() const override;
+        [[nodiscard]] double getShininess() const override;
 
-        [[nodiscard]] float getReflectivity() const override;
+        [[nodiscard]] double getReflectivity() const override;
 
     private:
         Component::Color _color1;
         Component::Color _color2;
-        float _squareSize;
-        float _reflectivity;
-        float _diffuseFactor;
+        double _squareSize;
+        double _reflectivity;
+        double _diffuseFactor;
 
         [[nodiscard]] Component::Color computeColor(const Component::Vector3f &hitPoint, const Component::Vector3f &normal,
-                                                    const Component::Vector3f &lightDirection, float lightIntensity,
-                                                    const Component::Color &ambientLightColor, float ambientLightIntensity) const override;
+                                                    const Component::Vector3f &lightDirection, double lightIntensity,
+                                                    const Component::Color &ambientLightColor, double ambientLightIntensity) const override;
     };
 }
