+++ src/Raytracer/Entity/ACam.hpp
@@ -19,11 +19,11 @@ namespace Raytracer {
     class ACam : public ICam, public AEntity {
     public:
         ACam(const std::string &typeCam, const Component::Vector3f &position, const Component::Vector3f &rotation,
-             float field_of_view);
+             double field_of_view);
 
         ~ACam() override { std::cout << "ACam::~ACam" << std::endl; };
 
-        [[nodiscard]] float getFieldOfView() const final;
+        [[nodiscard]] double getFieldOfView() const final;
 
         [[nodiscard]] const Component::Vector3f &getRotation() const;
         void setRotation(const Component::Vector3f &rotation);
@@ -33,7 +33,7 @@ namespace Raytracer {
     private:
         std::string _typeCam;
         Component::Vector3f _rotation;
-        float _fieldOfView;
+        double _fieldOfView;
     };
 };
