+++ src/Raytracer/Entity/ICam.hpp
@@ -15,7 +15,7 @@ namespace Raytracer {
     public:
         virtual ~ICam() = default;
 
-        [[nodiscard]] virtual float getFieldOfView() const = 0;
+        [[nodiscard]] virtual double getFieldOfView() const = 0;
 
     protected:
     private:
