+++ src/Raytracer/Matrix.hpp
@@ -17,12 +17,12 @@
 namespace Component {
 
     struct Matrix3x3 {
-        float m[9]{};
+        double m[9]{};
 
         Matrix3x3();
-        Matrix3x3(float m0, float m1, float m2,
-                  float m3, float m4, float m5,
-                  float m6, float m7, float m8);
+        Matrix3x3(double m0, double m1, double m2,
+                  double m3, double m4, double m5,
+                  double m6, double m7, double m8);
 
         static Matrix3x3 fromEulerAngles(const Component::Vector3f &angles);
         Matrix3x3 operator*(const Matrix3x3 &other) const;
